pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dj-rara"
version = "0.1.1"
description = "A lo-fi Spotify recommendation TUI — personalized tracks, mood-based discovery, playlist curation."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "spotipy>=2.23.0",
    "python-dotenv>=1.0.0",
    "textual>=8.2.2",
]

[project.scripts]
dj-rara = "dj_rara.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["dj_rara*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
